Feature/build opts

+++ flipside_platform/cmd/build.py
@@ -9,7 +9,7 @@ import os
 from .. import config
 
 
-def do_build():
+def do_build(**opts):
     app_config = config.get_app_config()
     build_dir = config.get_app_build_dir()
     build = app_config.get('build')
@@ -19,17 +19,22 @@ def do_build():
     module = importlib.import_module(module)
     fun = getattr(module, fun)
 
-    # copy salt deploy info into build dir
-    salt_dst_dir = os.path.join(build_dir,
-                                os.path.basename(config.get_salt_path()))
-    shutil.rmtree(salt_dst_dir, ignore_errors=True)
-    shutil.copytree(config.get_salt_path(), salt_dst_dir)
-
-    fun(build_dir, **build.get('function_kwargs', {}))
+    # XXX salt config (unused for now)
+    if os.path.exists(config.get_salt_path()):
+        salt_dst_dir = os.path.join(build_dir,
+                                    os.path.basename(config.get_salt_path()))
+        shutil.rmtree(salt_dst_dir, ignore_errors=True)
+        shutil.copytree(config.get_salt_path(), salt_dst_dir)
+    opts.update(build.get('function_kwargs', {}))
+    fun(build_dir, **opts)
 
 
 def main():
-    do_build()
+    import argparse
+    parser = argparse.ArgumentParser()
+    parser.add_argument('--environment', help='deploy environment (staging, production, ...)')
+    args = parser.parse_args()
+    do_build(**vars(args))
 
 if __name__ == '__main__':
     main()
